requirements.txt
Flask==0.12
requests==2.14.2
Keras
numpy
Pillow==4.0.0
scipy==0.18.1
tensorflow
h5py
theano==0.10.0beta3
Werkzeug==0.11.15
gunicorn==19.6.0
